Split out TestRun from ci/qunit-selenium-runner to separate file
This makes it easy to separate code that is not ours and include the
license at the top and document it's origin on the class.
